refactor build_watershed functions and re-arrange Pipfile
