+++ ce/api/streamflow/watershed.py
@@ -14,13 +14,10 @@ Functions `lonlat_to_xy()` and `xy_to_lonlat()`, which translate from a
 spatial tuple to a data index tuple and vice versa, also switch the
 dimension order accordingly.
 """
-from netCDF4 import Dataset
-import numpy as np
 import math
 from contexttimer import Timer
 
 from flask import abort
-from sqlalchemy import distinct
 from shapely.geometry import Point
 from shapely.errors import WKTReadingError
 from pint import UnitRegistry
@@ -31,14 +28,13 @@ from ce.api.geospatial import (
     WKT_point_to_lonlat,
     GeospatialTypeError,
 )
-from ce.api.util import is_valid_index, vec_add, neighbours
+from ce.api.util import neighbours
 from ce.geo_data_grid_2d import GeoDataGrid2DIndexError
 from ce.geo_data_grid_2d.vic import VicDataGrid
-from modelmeta import (
-    DataFile,
-    DataFileVariableGridded,
-    Ensemble,
-    EnsembleDataFileVariables,
+from ce.api.streamflow.shared import (
+    is_upstream,
+    VIC_direction_matrix,
+    get_time_invariant_variable_dataset,
 )
 
 
@@ -284,18 +280,6 @@ def build_watershed(target, routing, direction_map, debug=False):
     """
     visited = set()
 
-    def is_upstream(neighbour, cell):
-        """Return a boolean indicating whether `neighbour` is upstream of `cell`
-        according to the routing matrix and direction map."""
-        # Eliminate invalid cases
-        if not is_valid_index(neighbour, routing.shape):
-            return False
-        neighbour_routing = routing[neighbour]
-        if neighbour_routing is np.ma.masked:
-            return False
-        # `neighbour` is upstream if its routing points back at `cell`
-        return vec_add(neighbour, direction_map[int(neighbour_routing)]) == cell
-
     def upstream(cell):
         """Return all cells upstream of `cell`.
         This is the closure of upstream over cell neighbours.
@@ -306,73 +290,12 @@ def build_watershed(target, routing, direction_map, debug=False):
             *(
                 upstream(neighbour)
                 for neighbour in neighbours(cell)
-                if neighbour not in visited and is_upstream(neighbour, cell)
+                if neighbour not in visited
+                and is_upstream(neighbour, cell, routing, direction_map)
             )
         )
 
     return upstream(target)
-
-
-def VIC_direction_matrix(lat_step, lon_step):
-    """ Return a VIC direction matrix, which is a matrix indexed by the VIC
-    streamflow direction codes 0...9, with the value at index `i` indicating
-    the offsets from the data index in a streamflow file required to
-    step in that streamflow direction.
-
-    :param lat_step: Difference between two successive latitudes in streamflow
-        file. (Only) sign matters.
-    :param lon_step: Difference between two successive longitudes in streamflow
-        file. (Only) sign matters.
-    :return: tuple of offset pairs
-
-    The offsets must account for the sign of the step in the lat and lon
-    dimensions in the streamflow file.
-    For example, in a streamflow file with lat and lon both increasing with
-    increasing index, the offset to step northeast is [1, 1].
-
-    Note that argument order is (lat, lon), not (lon, lat).
-    """
-    base = (
-        (0, 0),  # filler - 0 is not used in the encoding
-        (1, 0),  # 1 = north
-        (1, 1),  # 2 = northeast
-        (0, 1),  # 3 = east
-        (-1, 1),  # 4 = southeast
-        (-1, 0),  # 5 = south
-        (-1, -1),  # 6 = southwest
-        (0, -1),  # 7 = west
-        (1, -1),  # 8 = northwest
-        (0, 0),  # 9 = outlet
-    )
-    lat_dir = int(math.copysign(1, lat_step))
-    lon_dir = int(math.copysign(1, lon_step))
-    return tuple(
-        (lat_dir * lat_base, lon_dir * lon_base) for lat_base, lon_base in base
-    )
-
-
-def get_time_invariant_variable_dataset(sesh, ensemble_name, variable):
-    """Locates and opens a time-invariant dataset.
-    These datasets contain things like elevation or area of a grid cell -
-    they're independent of time and there should be only one per ensemble.
-    If more or less than one is found in the ensemble, it raises an error
-
-    :param sesh: (sqlalchemy.orm.session.Session) A database Session object
-    :param ensemble_name: Name of the ensemble containing data files
-    :param variable: Name of *variable* inside dataset. This is not the
-        dataset filename or unique id.
-    :return: (netcdf.Dataset) netcdf Dataset object representing the file.
-    """
-    query = (
-        sesh.query(distinct(DataFile.filename).label("filename"))
-        .join(DataFileVariableGridded, EnsembleDataFileVariables, Ensemble,)
-        .filter(Ensemble.name == ensemble_name)
-        .filter(DataFileVariableGridded.netcdf_variable_name == variable)
-        .filter(DataFile.time_set_id.is_(None))
-    )
-
-    file = query.one()  # Raises exception if n != 1 results found
-    return Dataset(file.filename, "r")
 
 
 def hypsometry(elevations, areas, bin_start=0, bin_width=100, num_bins=46):
